Fixes segfault due to RenderProject copy constructor not being implemented
A copy of the RenderProject was being passed to the RenderThread. What that copy would get destructed, it would delete track_ member variable. That caused problems because it wasn't delete it's own version of that variable. Since I didn't want to support copying the RenderProject right now, I just pass a RenderProject pointer to the RenderThread instead.