ModifiableDrawObject's track redraw events independant of modification events.
We needed this because sometimes an object is modified in a way that just requires a redraw, but not necessarily a save (ex. bounding box visibility on RenderedObject). The inverse is also possible where an object is modified in a way where a save is required, but not necessarily a redraw (ex. a RenderEntities's name is changed).