2nd pass at scrub logic
mostly working, but it seems to miss the very last seek update. this is noticable if you scrub a large distance, and then perform a manual 1 step seek with the arrow buttons. watch the scrub bar, it will jump back to an older position, indicating the scrub bar and the GroupSeekerer were out of sync